GUI: fontsize,  kinetic toggle and other improvements (#527)
Gui changes:
 - Added kinetic scrolling to switch on the fly (fix #519)
 - Base font size can be changed in the GUI (fix #424 and fix #523)
 - Hide font selection by default, but allow it with config.ini entry. (should fix #523)

Further:
- Split GUI font size from screen font size
- GUI changes: move color buttons to the left of labels.
- Oculars: Separate font scaling for the Oculars GUI panel.
- Oculars: Storing of current instrument (telescope/ocular/CCD/lens) indices
- removed doubled parenting code in GUI classes

+++ src/core/modules/StarMgr.hpp
@@ -201,7 +201,7 @@ public slots:
 	double getLabelsAmount(void) const {return labelsAmount;}
 
 	//! Define font size to use for star names display.
-	void setFontSize(float newFontSize);
+	void setFontSize(int newFontSize);
 
 	//! Show scientific or catalog names on stars without common names.
 	static void setFlagSciNames(bool f) {flagSciNames = f;}
